requirements/base.txt
Django==3.2.3
django-allauth==0.50.0
django-autoslug==1.9.8
django-constance[database]
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-bulk==0.2.1
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.2.0
djangorestframework-gis==1.0
drf-nested-forms==1.1.7
drf-writable-nested==0.6.3
django-model-utils==4.1.1
django-rosetta==0.9.8
django-cors-headers==3.13.0
django-uuidfield==0.5.0
dj-rest-auth==2.2.4
django-select2==7.2.3

coreapi==2.3.3
Pillow==9.1.1
psycopg2-binary==2.9.3
redis==4.3.3
python-decouple==3.6


unidecode==1.2.0

# Images
sorl-thumbnail==12.8.0
sorl-thumbnail-serializer-field==0.2.1

# Tasks and jobs
django-redis==5.2.0

# deploy
h11==0.13.0
uvicorn==0.17.6
uvloop==0.16.0